backend-flask-server/requirements.txt
flask
gunicorn
torch
psycopg2-binary
transformers
python-dotenv
